tabla vehiculo

+++ RentCar.Models/Vehiculo.cs
@@ -15,7 +15,8 @@ namespace RentCar.Models
         public int CapacidadPersonas { get; set; }
         public string CapacidadMaletero { get; set; }
         //public string Categoria { get; set; }
-        public decimal Precio { get; set; }
+
+        public int Precio  { get; set; }
         public byte[] Foto { get; set; }
 
     }
